fix(setup): repair flow-registration paths + single source of truth (bd-1691)
register-all-flows.js pointed FLOW_CONFIGS at ../../shared/flows/*.json, but
the flow JSON lives in docs/flows/ — so the registrar crashed with ENOENT
before any Meta call. Extract FLOW_CONFIGS to flow-configs.js (correct
docs/flows/ paths incl. reading-assessment-flow-v2.json); register-all-flows
re-exports it. New test asserts every flow JSON path exists and parses.

Closes: bd-1691
